Improve docs for oci-distribution
