+++ crates/kubelet/src/image_client.rs
@@ -7,6 +7,27 @@ use oci_distribution::Reference;
 #[async_trait]
 pub trait ImageClient {
     /// Given a certain image reference pull the image data from the remote store
+    ///
+    /// # Example
+    /// ```rust
+    /// use async_trait::async_trait;
+    /// use kubelet::image_client::ImageClient;
+    /// use oci_distribution::Reference;
+    ///
+    /// struct InMemoryClient(std::collections::HashMap<Reference, Vec<u8>>);
+    ///
+    /// #[async_trait]
+    /// impl ImageClient for InMemoryClient {
+    ///     async fn pull(&mut self, image: &Reference) -> anyhow::Result<Vec<u8>> {
+    ///         let image = self
+    ///             .0
+    ///             .get(image)
+    ///             .clone()
+    ///             .ok_or(anyhow::anyhow!("Couldn't find image"))?;
+    ///         Ok(image.clone())
+    ///     }
+    /// }
+    /// ```
     async fn pull(&mut self, image: &Reference) -> anyhow::Result<Vec<u8>>;
 }
