+++ crates/oci-distribution/src/manifest.rs
@@ -1,3 +1,4 @@
+//! Oci Manifest
 use std::collections::HashMap;
 
 /// The mediatype for WASM layers.
